Revert "Push more logic into Screens services"
This reverts commit 633044844a51e3154aae771cdeaf3ecf25d38dd9.
