+++ app/helpers/steps_helper.rb
@@ -1,5 +1,29 @@
 module StepsHelper
+  def next_step_for(estimate, step)
+    next_estimate_step(steps_list_for(estimate), step)
+  end
+
   def previous_step_for(estimate, step)
-    Screens::PreviousScreenNameService.call(estimate, step)
+    next_estimate_step(steps_list_for(estimate).reverse, step)
+  end
+
+  def end_of_check_answer_loop?(estimate, current_step_name)
+    steps = StepListerService.call(estimate)
+    current_step = steps.find { _1.name == current_step_name }
+    return true if current_step.check_answer_group.nil?
+
+    current_index = steps.index(current_step)
+
+    steps[(current_index + 1)..].none? { _1.check_answer_group == current_step.check_answer_group }
+  end
+
+private
+
+  def steps_list_for(estimate)
+    StepListerService.call(estimate).map(&:name)
+  end
+
+  def next_estimate_step(steps, step)
+    steps.each_cons(2).detect { |old, _new| old == step }.last
   end
 end
